requirements.txt
pyaudio
websocket-client
flask
